test(Mix.Tasks.Ashes.Generate): tests for controller, model and channel
Tests have been added for the generation of controllers, models and
channels.

The tests are heavily based on:

  https://github.com/phoenixframework/phoenix/blob/ \
  3ae44468c5dfc89af61f549809e64c2f1b4ce95e/test/mix/tasks/phoenix/ \
  new_test.exs

Currently the tests do not clean up after each run. This means that the
test for creating a controller without the templates currently uses a
unique name to ensure it is not testing artifacts from a previous test.

Since ashes is intended to be used inside of a Phoenix project, the
`deps/ashes/templates` path is expected to exist. In the set up of the
tests, this folder is creates and the tests are copied manually to mimic
a real Phoenix application.